01_Crypto_App/requirements.txt
altair==5.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
frozendict==2.4.5
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
huggingface-hub==0.25.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
loguru==0.7.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
multitasking==0.0.11
narwhals==1.9.3
newsapi-python==0.2.7
nltk==3.9.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
peewee==3.17.6
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
protobuf==5.28.2
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
smmap==5.0.1
soupsieve==2.6
streamlit==1.39.0
tenacity==9.0.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tokenizers==0.20.1
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
webencodings==0.5.1
win32-setctime==1.1.0
wordcloud==1.9.3
yfinance==0.2.44
